panda: Implement placer and explorer MagicWords.
Note that these are quite buggy and are prone to crashing. Use with
caution.

Quote from Joey:

Make sure the models in {Panda3D}\models\misc are converted to .bam, or it
will crash.
To save you the trouble, you can get the converted models from here:
https://dl.dropboxusercontent.com/u/37515491/TTR/misc.zip - Tell someone
if the link is broken.