chat: Implement staff chat(s) using staff channels.
This was so much fun to write! Also, it doesn't matter if a hacked client
sends a different chatMode, because only staff will see it and they can
take action accordingly.

The message prefixing was only a temporary solution to let the other staff
members know that they are talking in a staff chat mode. Perhaps in the
future, this should use different chat bubble colours.